pyproject.toml
[tool.poetry]
name = "pm64r-discord-bot"
version = "0.1.0"
description = ""
authors = ["Pronyo <franktb93@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "==1.7.3"
"google-cloud-secret-manager" = "^2.10.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
